# openapi/sagewai.yaml
openapi: 3.1.0
info:
  title: Sagewai API
  version: 0.1.0
  description: REST API for Sagewai agent infrastructure platform.
  contact:
    email: hello@sagewai.ai
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: http://localhost:8100
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: API key or JWT token
  schemas:
    ChatMessage:
      type: object
      required: [role, content]
      properties:
        role: { type: string, enum: [system, user, assistant] }
        content: { type: string }
    ChatCompletionRequest:
      type: object
      required: [model, messages]
      properties:
        model: { type: string, example: gpt-4o }
        messages:
          type: array
          items: { $ref: '#/components/schemas/ChatMessage' }
        max_tokens: { type: integer }
        stream: { type: boolean, default: false }
        temperature: { type: number }
        tools: { type: array, items: { type: object } }
    ChatCompletionResponse:
      type: object
      properties:
        id: { type: string }
        object: { type: string, example: chat.completion }
        created: { type: integer }
        model: { type: string }
        choices:
          type: array
          items:
            type: object
            properties:
              index: { type: integer }
              message: { $ref: '#/components/schemas/ChatMessage' }
              finish_reason: { type: string }
        usage:
          type: object
          properties:
            prompt_tokens: { type: integer }
            completion_tokens: { type: integer }
            total_tokens: { type: integer }
...
paths:
...
  /v1/messages:
    post:
      operationId: createMessage
      tags: [Harness]
      summary: Anthropic Messages API (proxy)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [model, messages, max_tokens]
              properties:
                model: { type: string }
                messages: { type: array, items: { $ref: '#/components/schemas/ChatMessage' } }
                max_tokens: { type: integer }
                stream: { type: boolean, default: false }
                system: { type: string }
      responses:
        '200':
          description: Chat response or SSE stream
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
                  content: { type: array, items: { type: object } }
                  model: { type: string }
                  stop_reason: { type: string }
                  usage: { type: object }
  /v1/chat/completions:
    post:
      operationId: createChatCompletion
      tags: [Harness]
      summary: OpenAI Chat Completions API (proxy)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChatCompletionRequest' }
      responses:
        '200':
          description: Chat completion response or SSE stream
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChatCompletionResponse' }
...